LayoutTests/canvas/philip/tests/toDataURL.zerosize.html

<!DOCTYPE html>
<title>Canvas test: toDataURL.zerosize</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="0" height="0"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL();
_assertSame(data, 'data:,', "data", "'data:,'");


});
</script>
