LayoutTests/canvas/philip/tests/toDataURL.zerosize-expected.txt

< [index] >
toDataURL.zerosize
toDataURL on zero-size canvas returns 'data:,'
References: toDataURL.zerosize
Actual output:
Passed
